feat: add utility to correctly populate searchParams on fetch urls
