+++ frontend/src/utils/index.ts
@@ -1,0 +1,3 @@
+import { getNonEmptySearchParams } from './searchParams'
+
+export { getNonEmptySearchParams }
